CI: reorganized tests and markers
