+++ .github/workflows/tests_deb.yml
@@ -51,5 +51,5 @@ jobs:
         PYTEST_ADDOPTS: --durations=0 -vv --color=yes
       run: |
         sudo apt install -y nftables iptables ipset
-        pipenv run pytest -m deb_build
-        sudo -E pipenv run pytest -m deb_install
+        pipenv run pytest test/pkg/test_build_deb.py
+        sudo -E pipenv run pytest -m deb ./test/install
