+++ Pipfile
@@ -1,5 +1,5 @@
 [packages]
-pytest-cs = {ref = "0.7.5", git = "https://github.com/crowdsecurity/pytest-cs.git"}
+pytest-cs = {ref = "0.7.13", git = "https://github.com/crowdsecurity/pytest-cs.git"}
 pytest-dotenv = "0.5.2"
 pytest-dependency = "0.5.1"
 pexpect = "4.8.0"
