+++ pytest.ini
@@ -1,16 +1,12 @@
 [pytest]
 addopts =
     --pdbcls=IPython.terminal.debugger:Pdb
-    --ignore=rpm
+    --ignore=test/install
+    --ignore=test/backends
     --strict-markers
-    -m "not (deb_build or deb_install or rpm or scripts or iptables or nftables)"
 markers:
-    deb_build: tests for .deb packaging (non-root)
-    deb_install: tests for .deb packaging (root)
-    rpm: tests for .rpm packaging
-    scripts: test installation scripts (root)
-    iptables: tests for mode=iptables (root)
-    nftables: tests for mode=nftables (root)
+    deb: mark tests related to deb packaging
+    rpm: mark tests related to rpm packaging
 env_files =
     .env
     default.env
