+++ test/pkg/test_scripts_nonroot.py
